🐛 Fix hyfetch formatting by properly ordering .zshrc
Moved hyfetch call to the beginning of .zshrc above the p10k instant prompt to fix the hyfetch text not being colored. The new configuration is, to my knowledge, the intended way to use *fetch applications with p10k. This also allows the .p10k.zsh instant prompt mode variable to be unset from quiet, as hyfetch printing will not create a warning.